Website/Trading/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
newsmodel~=0.28
requests~=2.31.0
finnhub-python~=2.4.19